WMS and map tile providers

Named list of external spatial layers. By default layers are read in from package configuration file at ./data-raw/json/LAYERS.json. This file may be manually edited to add new layers to the dashboard. All spatial formats supported by the Leaflet JS viewer are supported (e.g. raster tiles, WMS, GeoJSON, KML, rasters, etc.), aside from vector tiles.

LAYERS

Format

A named list, names are provider codes.

Details

The list is organized by data provider (e.g. Digital Earth Africa, ESRI World Atlas, FAO WaPOR, etc.).

Examples

LAYERS[[2]]
#> $label
#> [1] "FAO WaPOR GeoServer"
#> 
#> $url
#> [1] "https://io.apps.fao.org/geoserver/wms?"
#> 
#> $legend
#> [1] "https://io.apps.fao.org/geoserver/wms?request=GetLegendGraphic&format=image/png&layer=%s"
#> 
#> $attr
#> [1] "FAO WaPOR"
#> 
#> $layers
#> $layers$`Natural Earth`
#> [1] "RICCAR:gray_hr_sr_w_3"
#> 
#> $layers$`Blue Marble`
#> [1] "NASA:bluemarble_ls_sw_st"
#> 
#> $layers$`Hydrological Basins`
#> [1] "AQUAMAPS:hydrobasins_africa"
#> 
#> $layers$`Major Rivers`
#> [1] "AQUAMAPS:rivers_africa"
#> 
#> $layers$`Land Cover 30m`
#> [1] "WAPOR_2:l3_odn_lcc_d"
#> 
#> $layers$`Monthly Precipitation 5km`
#> [1] "WAPOR_2:l1_pcp_m"
#> 
#> $layers$`Actual ET`
#> [1] "WAPOR_2:l3_odn_aeti_d"
#> 
#> $layers$`Incremental ET`
#> [1] "WATER:niger_etincr"
#> 
#> $layers$`Drought Intensity`
#> [1] "ASIS:di_d"
#> 
#> $layers$`Monthly Precipitation Anomaly`
#> [1] "NMME:fmpa"
#> 
#>