Bootstrap 4 accordion

This is a modified version of bs4Dash::accordion().

accordionItem to be inserted in a accordion container. This is a modified version of bs4Dash::accordionItem().

Change the selected value of an accordion input on the client. Copied from bs4Dash::updateAccordion().

Usage

accordion(..., id, width = 12)

accordionItem(
  ...,
  title,
  status = NULL,
  collapsed = TRUE,
  solidHeader = TRUE,
  class = NULL,
  bg = NULL
)

updateAccordion(id, selected, session = getDefaultReactiveDomain())

Arguments

...

slot for accordionItem.

id

unique accordion id.

width

the width of the accordion.

title

Optional title.

status

The status of the item. This determines the item's background color. Valid statuses are defined as follows:

primary: #007bff .

secondary: #6c757d .

info: #17a2b8 .

success: #28a745 .

warning: #ffc107 .

danger: #dc3545 .

gray-dark: #343a40 .

gray: #adb5bd .

white: #fff .

indigo: #6610f2 .

lightblue: #3c8dbc .

navy: #001f3f .

purple: #605ca8 .

fuchsia: #f012be .

pink: #e83e8c .

maroon: #d81b60 .

orange: #ff851b .

lime: #01ff70 .

teal: #39cccc .

olive: #3d9970 .

collapsed

If TRUE, start collapsed. This must be used with collapsible=TRUE.

solidHeader

Should the header be shown with a solid color background?

class

additional classes to apply to the parent card element

bg

background color

selected

Index of the newly selected accordionItem.

session

Shiny session object.