WMS and map tile providers

Named list of external spatial layers. By default layers are read in from package configuration file at ./data-raw/json/LAYERS.json. This file may be manually edited to add new layers to the dashboard. All spatial formats supported by the Leaflet JS viewer are supported (e.g. raster tiles, WMS, GeoJSON, KML, rasters, etc.), aside from vector tiles.

LAYERS

Format

A named list, names are provider codes.

Details

The list is organized by data provider (e.g. Digital Earth Africa, ESRI World Atlas, FAO WaPOR, etc.).

Examples

LAYERS[[2]]
#> $url
#> $url[[1]]
#> [1] "https://io.apps.fao.org/geoserver/wms?"
#> 
#> 
#> $attr
#> $attr[[1]]
#> [1] "FAO WaPOR"
#> 
#> 
#> $layers
#> $layers$`Natural Earth`
#> $layers$`Natural Earth`[[1]]
#> [1] "RICCAR:gray_hr_sr_w_3"
#> 
#> 
#> $layers$`Blue Marble`
#> $layers$`Blue Marble`[[1]]
#> [1] "NASA:bluemarble_ls_sw_st"
#> 
#> 
#> $layers$`Hydrological Basins`
#> $layers$`Hydrological Basins`[[1]]
#> [1] "AQUAMAPS:hydrobasins_africa"
#> 
#> 
#> $layers$`Major Rivers`
#> $layers$`Major Rivers`[[1]]
#> [1] "AQUAMAPS:rivers_africa"
#> 
#> 
#> $layers$`Land Cover 30m`
#> $layers$`Land Cover 30m`[[1]]
#> [1] "WAPOR_2:l3_odn_lcc_d"
#> 
#> 
#> $layers$`Monthly Precipitation 5km`
#> $layers$`Monthly Precipitation 5km`[[1]]
#> [1] "WAPOR_2:l1_pcp_m"
#> 
#> 
#> $layers$`Actual ET`
#> $layers$`Actual ET`[[1]]
#> [1] "WAPOR_2:l3_odn_aeti_d"
#> 
#> 
#> $layers$`Incremental ET`
#> $layers$`Incremental ET`[[1]]
#> [1] "WATER:niger_etincr"
#> 
#> 
#> $layers$`Drought Intensity`
#> $layers$`Drought Intensity`[[1]]
#> [1] "ASIS:di_d"
#> 
#> 
#> $layers$`Monthly Precipitation Anomaly`
#> $layers$`Monthly Precipitation Anomaly`[[1]]
#> [1] "NMME:fmpa"
#> 
#> 
#>